requirements.txt
awsebcli==3.5.4
backports.ssl-match-hostname==3.4.0.2
blessed==1.9.5
botocore==1.2.10
cement==2.4.0
colorama==0.3.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
gunicorn==19.3.0
jmespath==0.9.0
oauthlib==1.0.1
pathspec==0.3.3
psycopg2==2.6.1
PyJWT==1.4.0
pystache==0.5.4
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.13
PyYAML==3.11
requests==2.6.2
requests-oauthlib==0.5.0
six==1.9.0
static==1.1.1
static3==0.6.1
texttable==0.8.3
wcwidth==0.1.5
websocket-client==0.32.0
wheel==0.24.0
